+++ src/services/file/file.service.js
@@ -4,6 +4,7 @@ const path = require('path')
 const { FsUtils, ProcessUtils, LogUtils } = require('../../shared/utils')
 const { ScriptNotExist } = require('./script-not-exist.error')
 const { BadScriptPermission } = require('./bad-script-permission.error')
+const { TargetFileAlreadyExist } = require('./target-file-already-exist.error')
 
 class FileService {
   get scripts() {
@@ -115,63 +116,84 @@ class FileService {
       })
   }
 
-  deployModule(moduleName, global = true) {
-    const files = this.modules
-      .filter(element => element.module === moduleName)
-      .reduce((files, element) => {
-        for (const file of element.settings) {
-          files.push(file)
+  deployLocalFile({ source, target, global: isGlobalFile }, force = false) {
+    const deployPrommise = Promise.resolve()
+
+    if (isGlobalFile) {
+      deployPrommise.then(_ => { throw new TypeError('Unable to deploy global file as a local one.') })
+    }
+
+    const dirname = path.dirname(target)
+    if (!FsUtils.fileExist(dirname)) {
+      deployPrommise.then(_ => FsUtils.mkdirp(dirname))
+    }
+
+    return deployPrommise
+      .then(_ => FsUtils.lstat(target))
+      .catch(error => {
+        if (error.code !== 'ENOENT') {
+          throw error
         }
 
-        return files
-      }, [])
-      .filter(element => element.global === global)
+        return null // No file exist at target. No stats data to provide
+      })
+      .then(fileStats => {
+        if (fileStats != null && force === false) {
+          throw new TargetFileAlreadyExist(target)
+        }
 
-    const dirCreation = files
-      .map(file => path.dirname(file.target))
-      .filter(dir => !FsUtils.fileExist(dir))
-      .map(dir => FsUtils.mkdirp(dir))
+        return FsUtils.copyFile(source, target)
+      })
+      .catch(error => {
+        if (error.code !== 'EEXIST') {
+          throw error
+        }
 
-    return Promise.all(dirCreation).then(() => {
-      const linkCreation = files
-        .map(file => Promise.resolve()
-          .then(_ => FsUtils.lstat(file.target))
-          .then(targetStat => {
-            if (targetStat.isSymbolicLink()) {
-              return FsUtils.readlink(file.target)
-                .then(source => {
-                  if (source === file.source) {
-                    return FsUtils.unlink(file.target)
-                  }
-                })
-            }
-
-            if (targetStat.isFile() || targetStat.isDirectory()) {
-              return FsUtils.rename(file.target, `${file.target}.old`)
-            }
-          })
-          .catch(error => {
-            if (error.code !== 'ENOENT') { // No file exist at file.target
-              throw error
-            }
-          })
-          .then(_ => FsUtils.symlink(file.source, file.target))
-          .catch(error => {
-            if (error.code !== 'EEXIST') {
-              throw error
-            }
-
-            LogUtils.log({ type: 'warn', message: `Unable to link "${file.source}" => "${file.target}".` })
-          }))
-
-      return Promise.all(linkCreation)
-    })
+        LogUtils.log({ type: 'warn', message: `Unable to make a local copy ("${source}" => "${target}").` })
+      })
   }
 
-  deployModules(moduleNames, global = true) {
-    const modulesPromises = moduleNames.map(moduleName => this.deployModule(moduleName, global))
+  deployGlobalFile({ source, target, global: isGlobalFile }) {
+    const deployPrommise = Promise.resolve()
 
-    return Promise.all(modulesPromises)
+    if (!isGlobalFile) {
+      deployPrommise.then(_ => { throw new TypeError('Unable to deploy local file as a global one.') })
+    }
+
+    const dirname = path.dirname(target)
+    if (!FsUtils.fileExist(dirname)) {
+      deployPrommise.then(_ => FsUtils.mkdirp(dirname))
+    }
+
+    return deployPrommise
+      .then(_ => FsUtils.lstat(target))
+      .then(targetStat => {
+        if (targetStat.isSymbolicLink()) {
+          return FsUtils.readlink(target)
+            .then(linkSource => {
+              if (linkSource === source) {
+                return FsUtils.unlink(target)
+              }
+            })
+        }
+
+        if (targetStat.isFile() || targetStat.isDirectory()) {
+          return FsUtils.rename(target, `${target}.old`)
+        }
+      })
+      .catch(error => {
+        if (error.code !== 'ENOENT') { // No file exist at file.target
+          throw error
+        }
+      })
+      .then(_ => FsUtils.symlink(source, target))
+      .catch(error => {
+        if (error.code !== 'EEXIST') {
+          throw error
+        }
+
+        LogUtils.log({ type: 'warn', message: `Unable to link "${source}" => "${target}".` })
+      })
   }
 }