+++ src/services/file/target-file-already-exist.error.js
@@ -1,0 +1,9 @@
+class TargetFileAlreadyExist extends Error {
+  constructor(filename, message = `Target file "${filename}" already exist.`) {
+    super(message)
+
+    this.filename = filename
+  }
+}
+
+module.exports = exports = { TargetFileAlreadyExist }